Add env var to set max .wal size

+++ adapters/repos/db/lsmkv/bucket_options.go
@@ -46,9 +46,9 @@ func WithMinMMapSize(minMMapSize int64) BucketOption {
 	}
 }
 
-func WithMinWalThreshold(threshold uint64) BucketOption {
+func WithMinWalThreshold(threshold int64) BucketOption {
 	return func(b *Bucket) error {
-		b.minWalThreshold = threshold
+		b.minWalThreshold = uint64(threshold)
 		return nil
 	}
 }
